move some other dataclasses out from kube-client
