+++ platform_api/orchestrator/job_request.py
@@ -1,15 +1,13 @@
 import enum
-import shlex
 import uuid
-from dataclasses import asdict, dataclass, field
+from dataclasses import dataclass
 from pathlib import PurePath
 from typing import Any, Dict, List, Optional
 from urllib.parse import urlsplit
 
 from yarl import URL
 
-from platform_api.config import RegistryConfig
-from platform_api.resource import ResourcePoolType
+from platform_api.orchestrator.kube_client import Container, ContainerVolume
 
 
 class JobException(Exception):
@@ -22,202 +20,6 @@ class JobError(JobException):
 
 class JobNotFoundException(JobException):
     pass
-
-
-@dataclass(frozen=True)
-class ContainerVolume:
-    uri: URL
-    src_path: PurePath
-    dst_path: PurePath
-    read_only: bool = False
-
-    @staticmethod
-    def create(*args: Any, **kwargs: Any) -> "ContainerVolume":
-        return ContainerVolumeFactory(*args, **kwargs).create()
-
-    @classmethod
-    def from_primitive(cls, payload: Dict[str, Any]) -> "ContainerVolume":
-        kwargs = payload.copy()
-        # use dct.get() for backward compatibility
-        # old DB records has no src_uri field
-        kwargs["uri"] = URL(kwargs.get("uri", ""))
-        kwargs["src_path"] = PurePath(kwargs["src_path"])
-        kwargs["dst_path"] = PurePath(kwargs["dst_path"])
-        return cls(**kwargs)
-
-    def to_primitive(self) -> Dict[str, Any]:
-        payload: Dict[str, Any] = asdict(self)
-        payload["uri"] = str(payload["uri"])
-        payload["src_path"] = str(payload["src_path"])
-        payload["dst_path"] = str(payload["dst_path"])
-        return payload
-
-
-@dataclass(frozen=True)
-class ContainerResources:
-    cpu: float
-    memory_mb: int
-    gpu: Optional[int] = None
-    gpu_model_id: Optional[str] = None
-    shm: Optional[bool] = None
-
-    @classmethod
-    def from_primitive(cls, payload: Dict[str, Any]) -> "ContainerResources":
-        return cls(
-            cpu=payload["cpu"],
-            memory_mb=payload["memory_mb"],
-            gpu=payload.get("gpu"),
-            gpu_model_id=payload.get("gpu_model_id"),
-            shm=payload.get("shm"),
-        )
-
-    def to_primitive(self) -> Dict[str, Any]:
-        return asdict(self)
-
-    def check_fit_into_pool_type(self, pool_type: ResourcePoolType) -> bool:
-        if not self.gpu:
-            # container does not need GPU. we are good regardless of presence
-            # of GPU in the pool type.
-            return True
-
-        # container needs GPU
-
-        if not pool_type.gpu:
-            return False
-
-        if pool_type.gpu < self.gpu:
-            return False
-
-        if not self.gpu_model_id:
-            # container needs any GPU model
-            return True
-
-        assert pool_type.gpu_model
-        return self.gpu_model_id == pool_type.gpu_model.id
-
-
-@dataclass(frozen=True)
-class ContainerHTTPServer:
-    port: int
-    health_check_path: str = "/"
-    requires_auth: bool = False
-
-    @classmethod
-    def from_primitive(cls, payload: Dict[str, Any]) -> "ContainerHTTPServer":
-        return cls(
-            port=payload["port"],
-            health_check_path=payload.get("health_check_path") or cls.health_check_path,
-            requires_auth=payload.get("requires_auth", cls.requires_auth),
-        )
-
-    def to_primitive(self) -> Dict[str, Any]:
-        return asdict(self)
-
-
-@dataclass(frozen=True)
-class ContainerSSHServer:
-    port: int
-
-    @classmethod
-    def from_primitive(cls, payload: Dict[str, Any]) -> "ContainerSSHServer":
-        return cls(port=payload["port"])
-
-    def to_primitive(self) -> Dict[str, Any]:
-        return asdict(self)
-
-
-@dataclass(frozen=True)
-class Container:
-    image: str
-    resources: ContainerResources
-    command: Optional[str] = None
-    env: Dict[str, str] = field(default_factory=dict)
-    volumes: List[ContainerVolume] = field(default_factory=list)
-    http_server: Optional[ContainerHTTPServer] = None
-    ssh_server: Optional[ContainerSSHServer] = None
-
-    def belongs_to_registry(self, registry_config: RegistryConfig) -> bool:
-        prefix = f"{registry_config.host}/"
-        return self.image.startswith(prefix)
-
-    def to_image_uri(self, registry_config: RegistryConfig) -> URL:
-        assert self.belongs_to_registry(registry_config), "Unknown registry"
-        prefix = f"{registry_config.host}/"
-        repo = self.image.replace(prefix, "", 1)
-        uri = URL(f"image://{repo}")
-        path, *_ = uri.path.split(":", 1)
-        return uri.with_path(path)
-
-    @property
-    def port(self) -> Optional[int]:
-        if self.http_server:
-            return self.http_server.port
-        return None
-
-    @property
-    def ssh_port(self) -> Optional[int]:
-        if self.ssh_server:
-            return self.ssh_server.port
-        return None
-
-    @property
-    def health_check_path(self) -> str:
-        if self.http_server:
-            return self.http_server.health_check_path
-        return ContainerHTTPServer.health_check_path
-
-    @property
-    def command_list(self) -> List[str]:
-        if self.command:
-            return shlex.split(self.command)
-        return []
-
-    @property
-    def has_http_server_exposed(self) -> bool:
-        return bool(self.http_server)
-
-    @property
-    def has_ssh_server_exposed(self) -> bool:
-        return bool(self.ssh_server)
-
-    @property
-    def requires_http_auth(self) -> bool:
-        return bool(self.http_server and self.http_server.requires_auth)
-
-    @classmethod
-    def from_primitive(cls, payload: Dict[str, Any]) -> "Container":
-        kwargs = payload.copy()
-        kwargs["resources"] = ContainerResources.from_primitive(kwargs["resources"])
-        kwargs["volumes"] = [
-            ContainerVolume.from_primitive(item) for item in kwargs["volumes"]
-        ]
-
-        if kwargs.get("http_server"):
-            kwargs["http_server"] = ContainerHTTPServer.from_primitive(
-                kwargs["http_server"]
-            )
-        elif kwargs.get("port") is not None:
-            kwargs["http_server"] = ContainerHTTPServer.from_primitive(kwargs)
-
-        if kwargs.get("ssh_server"):
-            ssh_server_desc = kwargs["ssh_server"]
-            container_desc = ContainerSSHServer.from_primitive(ssh_server_desc)
-            kwargs["ssh_server"] = container_desc
-
-        kwargs.pop("port", None)
-        kwargs.pop("health_check_path", None)
-
-        return cls(**kwargs)
-
-    def to_primitive(self) -> Dict[str, Any]:
-        payload: Dict[str, Any] = asdict(self)
-        payload["resources"] = self.resources.to_primitive()
-        payload["volumes"] = [volume.to_primitive() for volume in self.volumes]
-        if self.http_server:
-            payload["http_server"] = self.http_server.to_primitive()
-        if self.ssh_server:
-            payload["ssh_server"] = self.ssh_server.to_primitive()
-        return payload
 
 
 @dataclass(frozen=True)
